pyproject.toml
[project]
name = "ai-chat-lib"
version = "0.1.0"
description = ""
authors = [
    {name = "Panda",email = "tgyezitang@163.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai (>=1.102.0,<2.0.0)",
    "google-genai (>=1.32.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "ai_chat_lib", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
